move: #17 storage api 분리
